+++ src/main/java/org/prgms/KdtApplication.java
@@ -1,0 +1,13 @@
+package org.prgms;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class KdtApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(KdtApplication.class, args);
+    }
+
+}
