+++ src/main/java/org/prgms/order/OrderRepository.java
@@ -1,7 +1,5 @@
 package org.prgms.order;
 
-import org.prgms.order.Order;
-
 public interface OrderRepository {
     public Order insert(Order order);
 }
