+++ src/test/java/org/prgms/OrderTester.java
@@ -1,0 +1,45 @@
+package org.prgms;
+
+import org.prgms.order.OrderItem;
+import org.prgms.order.OrderService;
+import org.prgms.voucher.FixedAmountVoucher;
+import org.prgms.voucher.VoucherRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import org.springframework.util.Assert;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.text.MessageFormat;
+import java.util.ArrayList;
+import java.util.UUID;
+
+public class OrderTester {
+    private static final Logger log = LoggerFactory.getLogger(OrderTester.class);
+
+    public static void main(String[] args) throws IOException {
+        var customerId = UUID.randomUUID();
+
+        var applicationContext = new AnnotationConfigApplicationContext(AppConfiguration.class);
+
+        var environment = applicationContext.getEnvironment();
+        var version = environment.getProperty("kdt.version");
+        var minimumOrderAmount = environment.getProperty("kdt.minimum-order-amount", Integer.class);
+
+        log.info("version -> {}", version);
+        log.info("minimumOrderAmount -> {}", minimumOrderAmount);
+
+        var voucherRepository = applicationContext.getBean(VoucherRepository.class);
+
+        var voucher = voucherRepository.save(new FixedAmountVoucher(UUID.randomUUID(), 10L));
+
+        var orderService = applicationContext.getBean(OrderService.class);
+
+        var order = orderService.createOrder(customerId, new ArrayList<OrderItem>() {{
+            add(new OrderItem(UUID.randomUUID(), 100L, 1));
+        }}, voucher.getVoucherId());
+        Assert.isTrue(order.totalAmount() == 90L,
+                MessageFormat.format("totalAmount({0}) is not 90L", order.totalAmount()));
+    }
+}
